Bug 786789 - Optimize TimeDuration::ToSeconds on Windows to avoid blocked store forwards; r=ehsan
UltraBlame original commit: 5502417dfd4e127f5de27e3a22857fd7af8ebb1e

+++ xpcom/ds/TimeStamp_windows.cpp
@@ -528,7 +528,8 @@ CalibratedPerformanceCounter()
 double
 TimeDuration::ToSeconds() const
 {
-  return double(mValue) / (sFrequencyPerSec * 1000ULL);
+  
+  return double(mValue) / (double(sFrequencyPerSec) * 1000.0);
 }
 
 double
